+++ src/server/cluster-gen/IClusterGenerator.ts
@@ -1,9 +1,0 @@
-import { IClusterData } from "@share/data/cluster-data";
-
-export interface IClusterGenerator {
-
-  /**
-   * Generate a standard world.
-   */
-  createWorldCluster(): IClusterData;
-}