+++ src/server/cluster-gen/biome-gen/IContinuousBiomeGenerator.ts
@@ -10,9 +10,4 @@ export interface IContinuousBiomeGenerator {
    * Given an x and z value, get the biomes and percentage compositions.
    */
   getBiomesFromXZ(x: number, z: number): IBiomeComposition;
-
-  /**
-   * Given an x and z value, get the top biome.
-   */
-  getTopBiomeFromXZ(x: number, z: number): Biome;
 }