+++ src/server/cluster-gen/biome-gen/index.ts
@@ -1,2 +1,3 @@
 export { ContinuousBiomeGenerator } from "./ContinuousBiomeGenerator";
+export { OneBiomeGenerator } from "./OneBiomeGenerator";
 export { IContinuousBiomeGenerator } from "./IContinuousBiomeGenerator";