+++ src/server/cluster-gen/index.ts
@@ -1,3 +1,1 @@
-export { DiscreteClusterGenerator } from "./DiscreteClusterGenerator";
 export { ContinuousClusterGenerator } from "./ContinuousClusterGenerator";
-export { IClusterGenerator } from "./IClusterGenerator";