+++ src/server/game-server/index.ts
@@ -1,2 +1,1 @@
 export { GameServer } from "./GameServer";
-export { IGameServer } from "./IGameServer";