+++ src/share/utility/Block.ts
@@ -1,5 +1,6 @@
 
 export enum Block {
+  Default,
   Air,
   Stone,
   Grass,