+++ changelog.md
@@ -1,5 +1,23 @@
+## Backlog
+* Send world data between client and server
+* Sound FX
+* Fix player movement behavior (going under arches, placing blocks under self, etc.)
+* Make asset manager static
+* Remove coordinate system and greatly simplify grid system
+
+## 0.0.4
+
+* Fixed lines between blocks
+* Improved top-down biome generation (inverse instead of logistic)
+
+
 ## 0.0.3
 
+* Add complex terrain generation with logistic curves
 * Add interpolation between biomes
 
 ## 0.0.2
+
+* First version.
+* Sinusoidal terrain generation
+* Block placing / breaking